+++ plugins/node/opentelemetry-instrumentation-pino/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.27.0](https://www.github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-pino-v0.26.0...instrumentation-pino-v0.27.0) (2021-11-19)
+
+
+### Features
+
+* **instrumentation-pino:** bump supported pino version to v7.x ([#724](https://www.github.com/open-telemetry/opentelemetry-js-contrib/issues/724)) ([aff24b8](https://www.github.com/open-telemetry/opentelemetry-js-contrib/commit/aff24b82e2654d6c42e1a9b03351dd94b9869793))
+
 ## [0.26.0](https://www.github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-pino-v0.25.0...instrumentation-pino-v0.26.0) (2021-10-22)